Fix connect test, which was passing an options object
An options object isn't supported, it was being used as an address, and
passed as a final argument to createConnection(port,...), where it was
ignored because it is not a function. This becomes noticeable once the
default port becomes correct, and you are running a STOMP server,
because this test ends up using a valid default port.