Document requirements for queue name
Also, fix the example. It didn't work against ActiveMQ, for which a
queue named 'foo' is considered to be the same as '/queue/foo', which is
fine, except that when the MESSAGE comes back on the queue, it uses the
fully-qualified destination, and stomp-client won't find a subscription
callback for that destination, so no messages were ever being printed.